requirements.txt
aiohttp==3.8.6       
aiosignal==1.3.1       
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.2
distro==1.8.0
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.335
langchain-experimental==0.0.40
langsmith==0.0.63
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.2.4
packaging==23.2
pandas==2.1.3
pip==22.3.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
yarl==1.9.2